update Multiple mode and alert

+++ app/cards/components/DeleteCard.js
@@ -1,4 +1,3 @@
-// app/cards/components/DeleteCard.js
 import React from "react";
 import { doc, deleteDoc } from "firebase/firestore";
 import { getStorage, ref, deleteObject } from "firebase/storage";
@@ -13,7 +12,11 @@ import {
 import { useState } from "react";
 import { useRouter } from "next/navigation";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faEllipsis, faPenToSquare, faTrashAlt } from "@fortawesome/free-solid-svg-icons";
+import {
+  faEllipsis,
+  faPenToSquare,
+  faTrashAlt,
+} from "@fortawesome/free-solid-svg-icons";
 
 const DeleteCard = ({
   deckId,
@@ -29,22 +32,31 @@ const DeleteCard = ({
   const storage = getStorage();
   const [isNavigating, setIsNavigating] = useState(false); // เพิ่ม state สำหรับการโหลดหน้า dashboard
   const router = useRouter();
+  const [loading, setLoading] = useState();
 
   const handleNavigate = () => {
     setIsNavigating(true); // ตั้งสถานะการโหลดเมื่อเริ่มเปลี่ยนเส้นทาง
-    router.push(`/EditCard/${cardId}?deckId=${deckId}`); // เปลี่ยนเส้นทางไปยังหน้า dashboard
+    router.push(`/EditCard/${cardId}?deckId=${deckId}`); // เปลี่ยนเส้นทางไปยังหน้า EditCard
   };
 
   const handleDelete = async () => {
     if (!auth) return;
+    setLoading(true);
+
+    const confirm = window.confirm("คุณแน่ใจหรือไม่ว่าต้องการลบการ์ด?");
+
+    if(!confirm){
+      setLoading(false);
+      return;
+    }
 
     const cardRef = doc(db, "Deck", auth.uid, "title", deckId, "cards", cardId);
 
     try {
-      // Delete the document
+      // ลบเอกสารใน Firestore
       await deleteDoc(cardRef);
 
-      // Delete images and audio from Firebase Storage
+      // ลบภาพและเสียงจาก Firebase Storage
       if (imageFront) {
         const imageFrontRef = ref(storage, imageFront);
         await deleteObject(imageFrontRef);
@@ -64,40 +76,42 @@ const DeleteCard = ({
         const audioBackRef = ref(storage, audioBack);
         await deleteObject(audioBackRef);
       }
-
-      console.log("Card and associated files deleted successfully");
+      console.log('Deleted card ')
+     
     } catch (error) {
       console.error("Error deleting card and files:", error);
     }
   };
 
   return (
-    <Dropdown>
-      <DropdownTrigger>
-        <div className="mt-2 cursor-pointer">
-          <FontAwesomeIcon
-            style={{ fontSize: "20px" }}
-            icon={faEllipsis}
-          ></FontAwesomeIcon>
-        </div>
-      </DropdownTrigger>
-      <DropdownMenu className=""  variant="flat">
-        <DropdownItem size="sm" color="warning" onClick={handleNavigate}>
-        <FontAwesomeIcon
-            style={{ fontSize: "20px" }}
-            icon={faPenToSquare}
-          ></FontAwesomeIcon>
-          EditCard
-        </DropdownItem>
-        <DropdownItem size="sm"  color="danger" onClick={handleDelete}>
-          <FontAwesomeIcon
-            style={{ fontSize: "20px" }}
-            icon={faTrashAlt}
-          ></FontAwesomeIcon>{" "}
-          Remove
-        </DropdownItem>
-      </DropdownMenu>
-    </Dropdown>
+    <>
+      <Dropdown>
+        <DropdownTrigger>
+          <div className="mt-2 cursor-pointer">
+            <FontAwesomeIcon
+              style={{ fontSize: "20px" }}
+              icon={faEllipsis}
+            ></FontAwesomeIcon>
+          </div>
+        </DropdownTrigger>
+        <DropdownMenu className="" variant="flat">
+          <DropdownItem size="sm" color="warning" onClick={handleNavigate}>
+            <FontAwesomeIcon
+              style={{ fontSize: "20px" }}
+              icon={faPenToSquare}
+            ></FontAwesomeIcon>
+            EditCard
+          </DropdownItem>
+          <DropdownItem size="sm" color="danger" onClick={handleDelete}> 
+            <FontAwesomeIcon
+              style={{ fontSize: "20px" }}
+              icon={faTrashAlt}
+            ></FontAwesomeIcon>{" "}
+            Remove
+          </DropdownItem>
+        </DropdownMenu>
+      </Dropdown>
+    </>
   );
 };
